+++ Library/Homebrew/dev-cmd/generate-formula-api.rb
@@ -50,6 +50,7 @@ module Homebrew
     FileUtils.mkdir_p directories
 
     Formulary.enable_factory_cache!
+    Formula.generating_hash!
 
     tap.formula_names.each do |name|
       formula = Formulary.factory(name)
