+++ Library/Homebrew/formula.rb
@@ -30,6 +30,7 @@ require "find"
 require "utils/spdx"
 require "extend/on_system"
 require "api"
+require "extend/api_hashable"
 
 # A formula provides instructions and metadata for Homebrew to install a piece
 # of software. Every Homebrew formula is a {Formula}.
@@ -69,6 +70,7 @@ class Formula
   extend Forwardable
   extend Cachable
   extend Predicable
+  extend APIHashable
 
   # The name of this {Formula}.
   # e.g. `this-formula`
