+++ Library/Homebrew/test/cask/cask_spec.rb
@@ -337,6 +337,8 @@ describe Cask::Cask, :cask do
       expect(JSON.pretty_generate(h["variations"])).to eq expected_sha256_variations.strip
     end
 
+    # @note The calls to `APIHashable.generating_hash!` and `APIHashable.generated_hash!`
+    #   are not idempotent so they can only be used in one test.
     it "returns the correct hash placeholders" do
       described_class.generating_hash!
       expect(described_class).to be_generating_hash
